refactor : 도메인별 서버/클라이언트 컴포넌트 구조 정리
