--- app/[universityName]/[lectureId]/components/index.ts
@@ -1,2 +1,0 @@
-export { default as LectureInfo } from './LectureInfo';
-export { default as ReviewList } from './ReviewList';
