--- app/[universityName]/[lectureId]/components/styles/LectureInfo.module.css
@@ -1,445 +1,0 @@
-.infoContainer {
-  border-radius: 20px;
-  border: none;
-  background-color: var(--white);
-  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.06);
-  margin: 2rem auto;
-  padding: 2rem;
-  transition: transform 0.3s ease;
-  width: 900px;
-}
-
-.infoContainer:hover {
-  transform: translateY(-4px) scale(1.01);
-  box-shadow: 0 12px 32px rgba(0, 0, 0, 0.12);
-}
-
-.introSection {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 2rem;
-  flex-wrap: wrap;
-}
-
-.leftContent {
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-  flex: 1 1 60%;
-  min-width: 260px;
-}
-
-.topLine {
-  width: 40px;
-  height: 4px;
-  background: linear-gradient(90deg, var(--primary), var(--yellow-500));
-  border-radius: 4px;
-}
-
-.lectureName {
-  font-size: 1.75rem;
-  font-weight: 700;
-  color: var(--text-primary);
-  margin: 0;
-  line-height: 1.3;
-}
-
-.professor {
-  font-size: 1.125rem;
-  color: var(--text-secondary);
-  margin-bottom: 0.25rem;
-}
-
-.infoBar {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.6rem;
-}
-
-.badge {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.35rem;
-  padding: 0.45rem 1rem;
-  border-radius: 9999px;
-  font-size: 0.8rem;
-  font-weight: 600;
-  background-color: var(--gray-50);
-  color: var(--text-primary);
-  border: 1px solid var(--gray-200);
-  max-width: 100%;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  transition: background 0.3s;
-}
-
-.opened {
-  background: var(--status-opened-bg);
-  color: var(--status-opened-text);
-  border: 1px solid var(--status-opened-border);
-}
-
-.closed {
-  background: var(--status-closed-bg);
-  color: var(--status-closed-text);
-  border: 1px solid var(--status-closed-border);
-}
-
-.icon {
-  stroke-width: 2;
-  flex-shrink: 0;
-}
-
-.introduction {
-  margin-top: 0.5rem;
-}
-
-.introduction p {
-  color: var(--text-secondary);
-  font-size: 0.95rem;
-  line-height: 1.5;
-  margin: 0;
-  word-break: break-word;
-}
-
-.statsSection {
-  display: flex;
-  align-items: center;
-  gap: 1.5rem;
-  margin-top: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid var(--gray-100);
-}
-
-.statItem {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.statIcon {
-  color: var(--primary);
-  stroke-width: 2;
-}
-
-.statNumber {
-  font-size: 1.25rem;
-  font-weight: 700;
-  color: var(--text-primary);
-}
-
-.statLabel {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-  font-weight: 500;
-}
-
-.rightContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  flex: 1 1 35%;
-  gap: 1.2rem;
-}
-
-.profileImage {
-  position: relative;
-  width: 160px;
-  height: 160px;
-  border-radius: 50%;
-  overflow: hidden;
-  box-shadow: 0 6px 20px rgba(0, 0, 0, 0.12);
-  background: linear-gradient(135deg, var(--yellow-100), var(--yellow-200));
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 50%;
-}
-
-@media (min-width: 1400px) {
-  .infoContainer {
-    width: 1000px;
-    padding: 2.5rem;
-  }
-
-  .lectureName {
-    font-size: 2rem;
-  }
-
-  .professor {
-    font-size: 1.25rem;
-  }
-
-  .profileImage {
-    width: 180px;
-    height: 180px;
-  }
-
-  .introduction p {
-    font-size: 1rem;
-  }
-}
-
-@media (max-width: 1200px) {
-  .infoContainer {
-    width: 800px;
-    padding: 2rem;
-  }
-
-  .lectureName {
-    font-size: 1.625rem;
-  }
-}
-
-@media (max-width: 1024px) {
-  .infoContainer {
-    width: 700px;
-    padding: 1.75rem;
-    margin: 1.5rem auto;
-  }
-
-  .introSection {
-    gap: 1.5rem;
-  }
-
-  .lectureName {
-    font-size: 1.5rem;
-  }
-
-  .profileImage {
-    width: 140px;
-    height: 140px;
-  }
-}
-
-@media (max-width: 900px) {
-  .infoContainer {
-    width: 600px;
-    padding: 1.5rem;
-  }
-
-  .introSection {
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-    gap: 1.5rem;
-  }
-
-  .leftContent {
-    flex: 1 1 100%;
-    align-items: center;
-  }
-
-  .rightContent {
-    flex: 1 1 100%;
-    flex-direction: row;
-    gap: 1.5rem;
-  }
-
-  .profileImage {
-    width: 120px;
-    height: 120px;
-  }
-
-  .infoBar {
-    justify-content: center;
-  }
-
-  .statsSection {
-    justify-content: center;
-  }
-}
-
-@media (max-width: 768px) {
-  .infoContainer {
-    width: 500px;
-    padding: 1.25rem;
-    margin: 1rem auto;
-    border-radius: 16px;
-  }
-
-  .introSection {
-    gap: 1.25rem;
-  }
-
-  .leftContent {
-    gap: 1rem;
-  }
-
-  .lectureName {
-    font-size: 1.375rem;
-  }
-
-  .professor {
-    font-size: 1rem;
-  }
-
-  .rightContent {
-    flex-direction: column;
-    gap: 0.75rem;
-  }
-
-  .profileImage {
-    width: 100px;
-    height: 100px;
-  }
-
-  .infoBar {
-    gap: 0.4rem;
-  }
-
-  .badge {
-    padding: 0.35rem 0.8rem;
-    font-size: 0.75rem;
-  }
-
-  .introduction p {
-    font-size: 0.9rem;
-  }
-
-  .statsSection {
-    gap: 1rem;
-  }
-
-  .statNumber {
-    font-size: 1.125rem;
-  }
-}
-
-@media (max-width: 600px) {
-  .infoContainer {
-    width: 450px;
-    padding: 1rem;
-  }
-
-  .lectureName {
-    font-size: 1.25rem;
-  }
-
-  .professor {
-    font-size: 0.9rem;
-  }
-
-  .profileImage {
-    width: 80px;
-    height: 80px;
-  }
-
-  .infoBar {
-    flex-direction: column;
-    align-items: center;
-    gap: 0.5rem;
-  }
-
-  .badge {
-    width: 100%;
-    justify-content: center;
-    max-width: 200px;
-  }
-
-  .statsSection {
-    flex-direction: column;
-    gap: 0.75rem;
-  }
-
-  .statItem {
-    justify-content: center;
-  }
-}
-
-@media (max-width: 480px) {
-  .infoContainer {
-    width: 100%;
-    max-width: 400px;
-    margin: 0.5rem auto;
-    padding: 1rem;
-    border-radius: 12px;
-  }
-
-  .introSection {
-    gap: 1rem;
-  }
-
-  .leftContent {
-    gap: 0.75rem;
-  }
-
-  .lectureName {
-    font-size: 1.125rem;
-  }
-
-  .professor {
-    font-size: 0.875rem;
-  }
-
-  .profileImage {
-    width: 70px;
-    height: 70px;
-  }
-
-  .topLine {
-    width: 32px;
-    height: 3px;
-  }
-
-  .infoBar {
-    gap: 0.4rem;
-  }
-
-  .badge {
-    padding: 0.3rem 0.7rem;
-    font-size: 0.7rem;
-    max-width: 180px;
-  }
-
-  .introduction p {
-    font-size: 0.85rem;
-  }
-
-  .statNumber {
-    font-size: 1rem;
-  }
-
-  .statLabel {
-    font-size: 0.8rem;
-  }
-}
-
-@media (max-width: 360px) {
-  .infoContainer {
-    padding: 0.75rem;
-    margin: 0.25rem auto;
-  }
-
-  .lectureName {
-    font-size: 1rem;
-  }
-
-  .professor {
-    font-size: 0.8rem;
-  }
-
-  .profileImage {
-    width: 60px;
-    height: 60px;
-  }
-
-  .badge {
-    padding: 0.25rem 0.6rem;
-    font-size: 0.65rem;
-    max-width: 160px;
-  }
-
-  .introduction p {
-    font-size: 0.8rem;
-  }
-}
