--- app/[universityName]/[lectureId]/components/styles/ReviewList.module.css
@@ -1,111 +1,0 @@
-.reviewContainer {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-.reviewHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 2rem;
-  padding-bottom: 1rem;
-  border-bottom: 1px solid var(--gray-100);
-}
-
-.reviewTitle {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: var(--text-primary);
-  margin: 0;
-}
-
-.reviewCount {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-  font-weight: 500;
-}
-
-.reviewList {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 1.5rem;
-}
-
-.emptyState {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  padding: 4rem 2rem;
-  text-align: center;
-  background: var(--white);
-  border-radius: 16px;
-  border: 1px solid var(--gray-100);
-  grid-column: 1 / -1;
-}
-
-.emptyIcon {
-  color: var(--text-secondary);
-  margin-bottom: 1rem;
-}
-
-.emptyTitle {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin: 0 0 0.5rem 0;
-}
-
-.emptyMessage {
-  font-size: 0.9rem;
-  color: var(--text-secondary);
-  margin: 0;
-  line-height: 1.5;
-}
-
-@media (max-width: 1024px) {
-  .reviewList {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 1.25rem;
-  }
-}
-
-@media (max-width: 768px) {
-  .reviewContainer {
-    padding: 1.5rem;
-  }
-
-  .reviewList {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-  }
-
-  .reviewHeader {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 0.5rem;
-  }
-
-  .reviewTitle {
-    font-size: 1.25rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .reviewContainer {
-    padding: 1rem;
-  }
-
-  .emptyState {
-    padding: 3rem 1.5rem;
-  }
-
-  .emptyTitle {
-    font-size: 1.125rem;
-  }
-
-  .emptyMessage {
-    font-size: 0.85rem;
-  }
-}
