--- app/[universityName]/components/index.ts
@@ -1,2 +1,0 @@
-export { default as LectureSelector } from './LectureSelector';
-export { default as LectureList } from './LectureList';
