--- app/[universityName]/components/styles/LectureSelector.module.css
@@ -1,89 +1,0 @@
-.lectureSelectorContainer {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  gap: 1rem;
-  /* border: 1px solid var(--gray-100); */
-  padding: 1rem 3.1rem;
-  /* border-top: solid 1px var(--gray-100);
-	border-left: solid 1px var(--gray-100);
-	border-right: solid 1px var(--gray-100); */
-	border-radius: 0.5rem 0.5rem 0 0;
-} 
-
-.lectureSelectorTitle {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
- 
-  h2 {
-    font-size: 1.2rem;
-    font-weight: 600;
-    color: var(--black);
-  }
-  
-}
-
-.lectureSelectorOptions {
-  display: flex;
-  flex-direction: row;
-  flex: 1;
-  justify-content: space-evenly;
-  padding: 0 10rem;
-}
-
-.lectureSelectorOption {
-	display: flex;
-	align-items: center;
-	border-radius: 4px;
-	overflow: hidden;
-	width: 200px;
-
-  span {
-    background-color: var(--yellow-200);
-    color: var(--white);
-    padding: 8px 12px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    white-space: nowrap;
-    width: 50%;
-  }
-
-  select {
-    padding: 8px 12px;
-    border: none;
-    background-color: var(--gray-300);
-    font-size: 14px;
-    color: var(--black);
-    height: 100%;
-    width: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-    appearance: none;
-    background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
-    background-position: right 0.5rem center;
-    background-repeat: no-repeat;
-    background-size: 1rem;
-    padding-right: 2rem;
-    transition: background-color 0.2s ease;
-  }
-
-  select:focus {
-    outline: none;
-  }
-
-  select:hover {
-    background-color: var(--gray-100);
-  }
-
-}
-
-.lectureSelectorImage {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
