--- app/providers/index.ts
@@ -1,1 +1,0 @@
-export { default as ReactQueryProvider } from './ReactQueryProvider';
