--- components/common/Selector/Selector.module.css
@@ -1,113 +1,0 @@
-.selectorContainer {
-  display: flex;
-  flex-direction:row;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.selectorTitle {
-  display: flex;
-  align-items: center;
-}
-
-.selectorTitle h2 {
-  font-size: 1.5rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin: 0;
-  position: relative;
-  padding-left: 1rem;
-}
-
-.selectorTitle h2::before {
-  content: '';
-  position: absolute;
-  left: 0;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 4px;
-  height: 20px;
-  background: linear-gradient(180deg, var(--primary) 0%, var(--yellow-500) 100%);
-  border-radius: 2px;
-}
-
-.selectorOptions {
-  display: flex;
-  gap: 1rem;
-  flex-wrap: wrap;
-  align-items: center;
-}
-
-.selectorOption {
-  display: flex;
-  align-items: center;
-  border-radius: 4px;
-  overflow: hidden;
-  width: 200px;
-  box-shadow: var(--shadow-sm);
-}
-
-.selectorOption span {
-  background-color: var(--yellow-200);
-  color: var(--black);
-  padding: 8px 12px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  white-space: nowrap;
-  width: 50%;
-  font-size: 0.875rem;
-  font-weight: 500;
-}
-
-.selectorOption select {
-  padding: 8px 12px;
-  border: none;
-  background-color: var(--gray-300);
-  font-size: 14px;
-  color: var(--black);
-  height: 100%;
-  width: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  cursor: pointer;
-  appearance: none;
-  background-image: url("data:image/svg+xml,%3csvg xmlns='http://www.w3.org/2000/svg' fill='none' viewBox='0 0 20 20'%3e%3cpath stroke='%236b7280' stroke-linecap='round' stroke-linejoin='round' stroke-width='1.5' d='M6 8l4 4 4-4'/%3e%3c/svg%3e");
-  background-position: right 0.5rem center;
-  background-repeat: no-repeat;
-  background-size: 1rem;
-  padding-right: 2rem;
-  transition: background-color 0.2s ease;
-}
-
-.selectorOption select:focus {
-  outline: none;
-}
-
-.selectorOption select:hover {
-  background-color: var(--gray-100);
-}
-
-.selectorImage {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-@media (max-width: 1080px) {
-  .selectorContainer {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 1rem;
-  }
-
-  .selectorOptions {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    gap: 1rem;
-  }
-
-}
