--- components/common/modal/LoginModal.module.css
@@ -1,51 +1,0 @@
-.loginForm {
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-  padding: 0.5rem 0.5rem;
-}
-
-.inputGroup {
-  display: flex;
-  flex-direction: column;
-  gap: 0.3rem;
-}
-
-.label {
-  font-size: 1rem;
-  font-weight: 500;
-  color: var(--text-primary);
-}
-
-.input {
-  padding: 0.7rem 1rem;
-  border: 1px solid var(--gray-200);
-  border-radius: 7px;
-  font-size: 1rem;
-  background-color: var(--gray-50);
-  color: var(--text-primary);
-  outline: none;
-  transition: border-color 0.2s;
-}
-
-.input:focus {
-  border-color: var(--gray-500);
-  background-color: var(--white);
-}
-
-.loginButton {
-  margin-top: 0.5rem;
-  padding: 0.7rem 0;
-  background-color: var(--primary);
-  color: var(--white);
-  border: none;
-  border-radius: 7px;
-  font-size: 1.05rem;
-  font-weight: 600;
-  cursor: pointer;
-  transition: background 0.18s;
-}
-
-.loginButton:hover {
-  background-color: var(--yellow-500);
-}
