--- components/common/modal/Modal.module.css
@@ -1,152 +1,0 @@
-.overlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000;
-  padding: 1rem;
-  animation: fadeIn 0.3s ease;
-}
-
-.modal {
-  background: var(--white);
-  border-radius: 16px;
-  box-shadow: 0 20px 60px rgba(0, 0, 0, 0.3);
-  max-height: 90vh;
-  overflow: hidden;
-  animation: slideIn 0.3s ease;
-  display: flex;
-  flex-direction: column;
-}
-
-.modal.small {
-  width: 100%;
-  max-width: 400px;
-}
-
-.modal.medium {
-  width: 100%;
-  max-width: 600px;
-}
-
-.modal.large {
-  width: 100%;
-  max-width: 800px;
-}
-
-.modalHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1.5rem 1.5rem 1rem 1.5rem;
-  border-bottom: 1px solid var(--gray-100);
-}
-
-.modalTitle {
-  font-size: 1.25rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin: 0;
-}
-
-.closeButton {
-  background: none;
-  border: none;
-  color: var(--text-secondary);
-  cursor: pointer;
-  padding: 0.5rem;
-  border-radius: 8px;
-  transition: all 0.2s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.closeButton:hover {
-  background: var(--gray-100);
-  color: var(--text-primary);
-}
-
-.modalContent {
-  padding: 1.5rem;
-  overflow-y: auto;
-  flex: 1;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
-
-@keyframes slideIn {
-  from {
-    opacity: 0;
-    transform: translateY(-20px) scale(0.95);
-  }
-  to {
-    opacity: 1;
-    transform: translateY(0) scale(1);
-  }
-}
-
-@media (max-width: 768px) {
-  .overlay {
-    padding: 0.5rem;
-  }
-
-  .modal {
-    max-height: 95vh;
-    border-radius: 12px;
-  }
-
-  .modalHeader {
-    padding: 1.25rem 1.25rem 0.75rem 1.25rem;
-  }
-
-  .modalTitle {
-    font-size: 1.125rem;
-  }
-
-  .modalContent {
-    padding: 1.25rem;
-  }
-
-  .modal.small {
-    max-width: 100%;
-  }
-
-  .modal.medium {
-    max-width: 100%;
-  }
-
-  .modal.large {
-    max-width: 100%;
-  }
-}
-
-@media (max-width: 480px) {
-  .overlay {
-    padding: 0.25rem;
-  }
-
-  .modalHeader {
-    padding: 1rem 1rem 0.5rem 1rem;
-  }
-
-  .modalTitle {
-    font-size: 1rem;
-  }
-
-  .modalContent {
-    padding: 1rem;
-  }
-}
