--- components/common/modal/ReviewDetailModal.module.css
@@ -1,194 +1,0 @@
-.reviewDetail {
-  display: flex;
-  flex-direction: column;
-  gap: 2rem;
-}
-
-.reviewHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  gap: 1rem;
-}
-
-.userInfo {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-}
-
-.profileImage {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 50%;
-}
-
-.userDetails {
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.username {
-  font-size: 1.125rem;
-  font-weight: 600;
-  color: var(--text-primary);
-}
-
-.reviewMeta {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.metaIcon {
-  color: var(--text-secondary);
-  stroke-width: 2;
-}
-
-.date {
-  font-size: 0.875rem;
-  color: var(--text-secondary);
-}
-
-.ratingSection {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.reviewContent {
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-}
-
-.reviewTitle {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: var(--text-primary);
-  margin: 0;
-  line-height: 1.3;
-}
-
-.reviewText {
-  display: flex;
-  flex-direction: column;
-  gap: 1rem;
-}
-
-.paragraph {
-  font-size: 1rem;
-  color: var(--text-secondary);
-  line-height: 1.7;
-  margin: 0;
-  word-break: break-word;
-}
-
-.reviewFooter {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  padding-top: 1.5rem;
-  border-top: 1px solid var(--gray-100);
-}
-
-.likesSection {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-  padding: 0.75rem 1rem;
-  border-radius: 24px;
-  background: var(--gray-50);
-  transition: background 0.3s ease;
-}
-
-.likesSection:hover {
-  background: var(--gray-100);
-}
-
-.likeIcon {
-  color: var(--text-secondary);
-  stroke-width: 2;
-}
-
-.likeCount {
-  font-size: 0.9rem;
-  color: var(--text-secondary);
-  font-weight: 500;
-}
-
-@media (max-width: 768px) {
-  .reviewDetail {
-    gap: 1.5rem;
-  }
-
-  .reviewHeader {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 1rem;
-  }
-
-  .ratingSection {
-    align-self: flex-end;
-  }
-
-  .reviewTitle {
-    font-size: 1.25rem;
-  }
-
-  .paragraph {
-    font-size: 0.95rem;
-  }
-
-  .username {
-    font-size: 1rem;
-  }
-
-  .profileImage {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-@media (max-width: 480px) {
-  .reviewDetail {
-    gap: 1.25rem;
-  }
-
-  .reviewTitle {
-    font-size: 1.125rem;
-  }
-
-  .paragraph {
-    font-size: 0.9rem;
-  }
-
-  .username {
-    font-size: 0.95rem;
-  }
-
-  .profileImage {
-    width: 36px;
-    height: 36px;
-  }
-
-  .likesSection {
-    padding: 0.6rem 0.8rem;
-  }
-
-  .likeCount {
-    font-size: 0.85rem;
-  }
-}
