--- components/common/starRating/StarRating.module.css
@@ -1,70 +1,0 @@
-.ratingContainer {
-  display: flex;
-  align-items: center;
-  gap: 6px;
-}
-
-.starRating {
-  display: flex;
-  align-items: center;
-  gap: 1px;
-}
-
-.star {
-  color: var(--star-color);
-  text-shadow: 0 1px 2px rgba(245, 158, 11, 0.3);
-}
-
-.halfStar {
-  color: var(--star-color);
-  opacity: 0.7;
-  text-shadow: 0 1px 2px rgba(245, 158, 11, 0.3);
-}
-
-.emptyStar {
-  color: var(--text-muted);
-  opacity: 0.5;
-}
-
-.ratingText {
-  padding-left: 0.4rem;
-  color: var(--text-secondary);
-  font-weight: 600;
-}
-
-.noRating {
-  color: var(--text-muted);
-  font-style: italic;
-}
-
-.small {
-  font-size: 12px;
-}
-
-.medium {
-  font-size: 16px;
-}
-
-.large {
-  font-size: 20px;
-}
-
-.veryLarge {
-  font-size: 24px;
-}
-
-.ratingText.small {
-  font-size: 10px;
-}
-
-.ratingText.medium {
-  font-size: 14px;
-}
-
-.ratingText.large {
-  font-size: 16px;
-} 
-
-.ratingText.veryLarge {
-  font-size: 18px;
-}
