--- components/ui/index.ts
@@ -1,4 +1,0 @@
-export { default as Header } from './header/Header';
-export { default as HamburgerMenu } from './hamburgerMenu/HamburgerMenu';
-export { default as UniversityList } from './universityList/UniversityList';
-export { default as DropdownUniversityList } from './universityList/DropdownUniversityList';
