--- components/ui/universityList/University.module.css
@@ -1,109 +1,0 @@
-.universityCard {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: 10px;
-  border-radius: 8px;
-  transition:
-    background-color 0.2s ease,
-    transform 0.2s cubic-bezier(0.4, 0, 0.2, 1);
-  will-change: background-color, transform;
-}
-
-.universityGrid {
-  display: grid;
-  justify-content: center;
-  align-items: center;
-  grid-template-columns: repeat(auto-fit, minmax(100px, 1fr));
-  gap: 10px;
-  width: 100%;
-}
-
-.universityGrid:hover .universityCard {
-  transform: scale(0.9);
-  z-index: 1;
-  filter: grayscale(1);
-}
-
-.universityGrid .universityCard:hover {
-  transform: scale(1.1);
-  z-index: 2;
-  filter: none;
-}
-
-.universityImage {
-  display: block;
-  max-width: 100%;
-  height: auto;
-  object-fit: contain;
-  pointer-events: none;
-}
-
-.centerDropdown {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-}
-
-.centerDropdown.hidden {
-  display: none;
-}
-
-.dropdownTrigger {
-  width: auto;
-  background: none;
-  border: none; 
-  font-size: 1.3rem;
-  font-weight: 500;
-  color: var(--black);
-  cursor: pointer;
-  transition: all 0.2s ease;
-  height: 100%;
-}
-
-.dropdownTrigger:hover {
-  color: var(--primary);
-}
-
-.dropdownMenu {
-  position: absolute;
-  top: 130%;
-  left: 50%;
-  transform: translateX(-50%);
-  background-color: var(--white);
-  border-radius: 12px;
-  box-shadow: 0 10px 40px rgba(0, 0, 0, 0.15);
-  padding: 1.5rem;
-  z-index: 1000;
-  margin-top: 0.5rem;
-  min-width: 900px;
-  opacity: 0;
-  pointer-events: none;
-  transition: opacity 0.1s ease;
-}
-
-.dropdownMenu.open {
-  opacity: 1;
-  pointer-events: auto;
-}
-
-@media (max-width: 1024px) {
-  .dropdownMenu {
-    min-width: 300px;
-  }
-}
-
-.dropdownMenu::before {
-  content: '';
-  position: absolute;
-  top: -12px;
-  left: 50%;
-  transform: translateX(-50%);
-  width: 0;
-  height: 0;
-  border-left: 12px solid transparent;
-  border-right: 12px solid transparent;
-  border-bottom: 12px solid var(--white);
-}
