+++ domains/lecture/index.ts
@@ -1,0 +1,5 @@
+// Lecture Domain - Server Components
+export { default as LectureCard } from './server/components/lecture-card';
+export { default as LectureList } from './server/components/lecture-list';
+export { default as LectureInfo } from './server/components/lecture-info';
+export { default as LectureSelector } from './server/components/lecture-selector';
