+++ domains/lecture/server/components/lecture-info.tsx
