+++ domains/lecture/server/components/lecture-list.tsx
@@ -1,6 +1,6 @@
 import { Lecture, LectureResult } from '@/types';
-import LectureCard from '@/domains/lecture/components/LectureCard';
-import styles from './styles/LectureList.module.css';
+import LectureCard from '@/components/server/cards/lecture-card';
+import styles from '@/app/[universityName]/components/styles/LectureList.module.css';
 import Link from 'next/link';
 
 interface LectureListProps {
