--- domains/lecture/skeleton/LectureInfoSkeleton.module.css
@@ -1,429 +1,0 @@
-@keyframes shimmer {
-  0% {
-    background-position: -200px 0;
-  }
-  100% {
-    background-position: calc(200px + 100%) 0;
-  }
-}
-
-.skeleton {
-  background: linear-gradient(
-    90deg,
-    #f0f0f0 25%, 
-    #e0e0e0 50%, 
-    #f0f0f0 75%
-  );
-  background-size: 200px 100%;
-  animation: shimmer 1.5s infinite;
-  border-radius: 4px;
-}
-
-.infoContainer {
-  border-radius: 20px;
-  border: none;
-  background-color: var(--white);
-  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.06);
-  margin: 2rem auto;
-  padding: 2rem;
-  width: 900px;
-}
-
-.introSection {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 2rem;
-  flex-wrap: wrap;
-}
-
-.leftContent {
-  display: flex;
-  flex-direction: column;
-  gap: 1.2rem;
-  flex: 1 1 60%;
-  min-width: 260px;
-}
-
-.topLine {
-  composes: skeleton;
-  width: 40px;
-  height: 4px;
-  border-radius: 4px;
-}
-
-.lectureNameSkeleton {
-  composes: skeleton;
-  height: 28px;
-  width: 80%;
-  border-radius: 6px;
-}
-
-.professorSkeleton {
-  composes: skeleton;
-  height: 18px;
-  width: 50%;
-  border-radius: 4px;
-}
-
-.infoBar {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 0.6rem;
-}
-
-.badgeSkeleton {
-  composes: skeleton;
-  height: 32px;
-  width: 100px;
-  border-radius: 9999px;
-}
-
-.introductionSkeleton {
-  margin-top: 0.5rem;
-  display: flex;
-  flex-direction: column;
-  gap: 0.5rem;
-}
-
-.introLineSkeleton {
-  composes: skeleton;
-  height: 16px;
-  border-radius: 4px;
-}
-
-.introLineSkeleton:nth-child(1) {
-  width: 100%;
-}
-
-.introLineSkeleton:nth-child(2) {
-  width: 85%;
-}
-
-.introLineSkeleton:nth-child(3) {
-  width: 70%;
-}
-
-.statsSection {
-  display: flex;
-  align-items: center;
-  gap: 1.5rem;
-  margin-top: 1rem;
-  padding-top: 1rem;
-  border-top: 1px solid var(--gray-100);
-}
-
-.statItem {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.statIconSkeleton {
-  composes: skeleton;
-  width: 16px;
-  height: 16px;
-  border-radius: 50%;
-}
-
-.statNumberSkeleton {
-  composes: skeleton;
-  height: 20px;
-  width: 30px;
-  border-radius: 4px;
-}
-
-.statLabelSkeleton {
-  composes: skeleton;
-  height: 14px;
-  width: 60px;
-  border-radius: 4px;
-}
-
-.starRatingSkeleton {
-  composes: skeleton;
-  height: 24px;
-  width: 120px;
-  border-radius: 12px;
-}
-
-.rightContent {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  flex: 1 1 35%;
-  gap: 1.2rem;
-}
-
-.profileImageSkeleton {
-  composes: skeleton;
-  width: 160px;
-  height: 160px;
-  border-radius: 50%;
-}
-
-@media (min-width: 1400px) {
-  .infoContainer {
-    width: 1000px;
-    padding: 2.5rem;
-  }
-
-  .lectureNameSkeleton {
-    height: 32px;
-  }
-
-  .professorSkeleton {
-    height: 20px;
-  }
-
-  .profileImageSkeleton {
-    width: 180px;
-    height: 180px;
-  }
-}
-
-@media (max-width: 1200px) {
-  .infoContainer {
-    width: 800px;
-    padding: 2rem;
-  }
-
-  .lectureNameSkeleton {
-    height: 26px;
-  }
-}
-
-@media (max-width: 1024px) {
-  .infoContainer {
-    width: 700px;
-    padding: 1.75rem;
-    margin: 1.5rem auto;
-  }
-
-  .introSection {
-    gap: 1.5rem;
-  }
-
-  .lectureNameSkeleton {
-    height: 24px;
-  }
-
-  .profileImageSkeleton {
-    width: 140px;
-    height: 140px;
-  }
-}
-
-@media (max-width: 900px) {
-  .infoContainer {
-    width: 600px;
-    padding: 1.5rem;
-  }
-
-  .introSection {
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-    gap: 1.5rem;
-  }
-
-  .leftContent {
-    flex: 1 1 100%;
-    align-items: center;
-  }
-
-  .rightContent {
-    flex: 1 1 100%;
-    flex-direction: row;
-    gap: 1.5rem;
-  }
-
-  .profileImageSkeleton {
-    width: 120px;
-    height: 120px;
-  }
-
-  .infoBar {
-    justify-content: center;
-  }
-
-  .statsSection {
-    justify-content: center;
-  }
-}
-
-@media (max-width: 768px) {
-  .infoContainer {
-    width: 500px;
-    padding: 1.25rem;
-    margin: 1rem auto;
-    border-radius: 16px;
-  }
-
-  .introSection {
-    gap: 1.25rem;
-  }
-
-  .leftContent {
-    gap: 1rem;
-  }
-
-  .lectureNameSkeleton {
-    height: 22px;
-  }
-
-  .professorSkeleton {
-    height: 16px;
-  }
-
-  .rightContent {
-    flex-direction: column;
-    gap: 0.75rem;
-  }
-
-  .profileImageSkeleton {
-    width: 100px;
-    height: 100px;
-  }
-
-  .infoBar {
-    gap: 0.4rem;
-  }
-
-  .badgeSkeleton {
-    height: 28px;
-    width: 80px;
-  }
-
-  .statsSection {
-    gap: 1rem;
-  }
-
-  .statNumberSkeleton {
-    height: 18px;
-    width: 25px;
-  }
-}
-
-@media (max-width: 600px) {
-  .infoContainer {
-    width: 450px;
-    padding: 1rem;
-  }
-
-  .lectureNameSkeleton {
-    height: 20px;
-  }
-
-  .professorSkeleton {
-    height: 14px;
-  }
-
-  .profileImageSkeleton {
-    width: 80px;
-    height: 80px;
-  }
-
-  .infoBar {
-    flex-direction: column;
-    align-items: center;
-    gap: 0.5rem;
-  }
-
-  .badgeSkeleton {
-    width: 100%;
-    max-width: 200px;
-  }
-
-  .statsSection {
-    flex-direction: column;
-    gap: 0.75rem;
-  }
-
-  .statItem {
-    justify-content: center;
-  }
-}
-
-@media (max-width: 480px) {
-  .infoContainer {
-    width: 100%;
-    max-width: 400px;
-    margin: 0.5rem auto;
-    padding: 1rem;
-    border-radius: 12px;
-  }
-
-  .introSection {
-    gap: 1rem;
-  }
-
-  .leftContent {
-    gap: 0.75rem;
-  }
-
-  .lectureNameSkeleton {
-    height: 18px;
-  }
-
-  .professorSkeleton {
-    height: 14px;
-  }
-
-  .profileImageSkeleton {
-    width: 70px;
-    height: 70px;
-  }
-
-  .topLine {
-    width: 32px;
-    height: 3px;
-  }
-
-  .infoBar {
-    gap: 0.4rem;
-  }
-
-  .badgeSkeleton {
-    height: 24px;
-    width: 70px;
-    max-width: 180px;
-  }
-
-  .statNumberSkeleton {
-    height: 16px;
-    width: 20px;
-  }
-
-  .statLabelSkeleton {
-    height: 12px;
-    width: 50px;
-  }
-}
-
-@media (max-width: 360px) {
-  .infoContainer {
-    padding: 0.75rem;
-    margin: 0.25rem auto;
-  }
-
-  .lectureNameSkeleton {
-    height: 16px;
-  }
-
-  .professorSkeleton {
-    height: 12px;
-  }
-
-  .profileImageSkeleton {
-    width: 60px;
-    height: 60px;
-  }
-
-  .badgeSkeleton {
-    height: 20px;
-    width: 60px;
-    max-width: 160px;
-  }
-}
