+++ domains/review/index.ts
@@ -1,0 +1,7 @@
+// Review Domain - Server Components
+export { default as ReviewCard } from './server/components/review-card';
+export { default as ReviewList } from './server/components/review-list';
+
+// Review Domain - Client Components
+export { default as ReviewCardModal } from './client/components/ReviewCardModal';
+export { default as ReviewDetailModal } from './client/components/ReviewDetailModal';
