--- domains/review/skeleton/ReviewCardListSkeleton.module.css
@@ -1,91 +1,0 @@
-@keyframes shimmer {
-  0% {
-    background-position: -200px 0;
-  }
-  100% {
-    background-position: calc(200px + 100%) 0;
-  }
-}
-
-.skeleton {
-  background: linear-gradient(
-    90deg,
-    #f0f0f0 25%, 
-    #e0e0e0 50%, 
-    #f0f0f0 75%
-  );
-  background-size: 200px 100%;
-  animation: shimmer 1.5s infinite;
-  border-radius: 4px;
-}
-
-.reviewContainer {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 2rem;
-}
-
-.reviewHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 2rem;
-  padding-bottom: 1rem;
-  border-bottom: 1px solid var(--gray-100);
-}
-
-.reviewTitleSkeleton {
-  composes: skeleton;
-  height: 24px;
-  width: 120px;
-}
-
-.reviewCountSkeleton {
-  composes: skeleton;
-  height: 16px;
-  width: 80px;
-}
-
-.reviewList {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 1.5rem;
-}
-
-@media (max-width: 1024px) {
-  .reviewList {
-    grid-template-columns: repeat(2, 1fr);
-    gap: 1.25rem;
-  }
-}
-
-@media (max-width: 768px) {
-  .reviewContainer {
-    padding: 1.5rem;
-  }
-
-  .reviewList {
-    grid-template-columns: 1fr;
-    gap: 1rem;
-  }
-
-  .reviewHeader {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 0.5rem;
-  }
-
-  .reviewTitleSkeleton {
-    height: 20px;
-  }
-}
-
-@media (max-width: 480px) {
-  .reviewContainer {
-    padding: 1rem;
-  }
-
-  .reviewTitleSkeleton {
-    height: 18px;
-  }
-}
