--- domains/review/styles/ReviewCard.module.css
@@ -1,214 +1,0 @@
-.reviewCard {
-  background: var(--white);
-  border-radius: 16px;
-  border: 1px solid var(--gray-100);
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.04);
-  padding: 1.25rem;
-  transition: all 0.3s ease;
-  display: flex;
-  flex-direction: column;
-  height: fit-content;
-  cursor: pointer;
-}
-
-.reviewCard:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.08);
-}
-
-.cardHeader {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  margin-bottom: 1rem;
-}
-
-.userInfo {
-  display: flex;
-  align-items: center;
-  gap: 0.75rem;
-}
-
-.profileImage {
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  overflow: hidden;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.image {
-  width: 100%;
-  height: 100%;
-  object-fit: cover;
-  border-radius: 50%;
-}
-
-.userDetails {
-  display: flex;
-  flex-direction: column;
-  gap: 0.25rem;
-}
-
-.username {
-  font-size: 0.9rem;
-  font-weight: 600;
-  color: var(--text-primary);
-}
-
-.reviewMeta {
-  display: flex;
-  align-items: center;
-  gap: 0.25rem;
-}
-
-.metaIcon {
-  color: var(--text-secondary);
-  stroke-width: 2;
-}
-
-.date {
-  font-size: 0.75rem;
-  color: var(--text-secondary);
-}
-
-.ratingSection {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.ratingText {
-  font-size: 0.875rem;
-  font-weight: 600;
-  color: var(--primary);
-}
-
-.cardBody {
-  margin-bottom: 1rem;
-  flex: 1;
-}
-
-.reviewTitle {
-  font-size: 1rem;
-  font-weight: 600;
-  color: var(--text-primary);
-  margin: 0 0 0.75rem 0;
-  line-height: 1.4;
-}
-
-.reviewContent {
-  font-size: 0.85rem;
-  color: var(--text-secondary);
-  line-height: 1.6;
-  margin: 0;
-  display: -webkit-box;
-  -webkit-line-clamp: 1;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.cardFooter {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  margin-top: auto;
-}
-
-.likesSection {
-  display: flex;
-  align-items: center;
-  gap: 0.25rem;
-  padding: 0.5rem 0.75rem;
-  border-radius: 20px;
-  background: var(--gray-50);
-  transition: background 0.3s ease;
-}
-
-.likesSection:hover {
-  background: var(--gray-100);
-}
-
-.likeIcon {
-  color: var(--text-secondary);
-  stroke-width: 2;
-}
-
-.likeCount {
-  font-size: 0.8rem;
-  color: var(--text-secondary);
-  font-weight: 500;
-}
-
-@media (max-width: 1024px) {
-  .reviewCard {
-    padding: 1.125rem;
-  }
-}
-
-@media (max-width: 768px) {
-  .reviewCard {
-    padding: 1.25rem;
-  }
-
-  .cardHeader {
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 1rem;
-  }
-
-  .ratingSection {
-    align-self: flex-end;
-  }
-
-  .reviewTitle {
-    font-size: 1rem;
-  }
-
-  .reviewContent {
-    font-size: 0.85rem;
-  }
-}
-
-@media (max-width: 480px) {
-  .reviewCard {
-    padding: 1rem;
-    border-radius: 12px;
-  }
-
-  .userInfo {
-    gap: 0.5rem;
-  }
-
-  .profileImage {
-    width: 32px;
-    height: 32px;
-  }
-
-  .username {
-    font-size: 0.85rem;
-  }
-
-  .date {
-    font-size: 0.7rem;
-  }
-
-  .reviewTitle {
-    font-size: 0.95rem;
-  }
-
-  .reviewContent {
-    font-size: 0.8rem;
-  }
-
-  .likesSection {
-    padding: 0.4rem 0.6rem;
-  }
-
-  .likeCount {
-    font-size: 0.75rem;
-  }
-}
