+++ shared/components/index.ts
@@ -1,0 +1,17 @@
+// Shared Server Components
+export { default as StarRating } from './server/star-rating';
+export { default as Selector } from './server/selector';
+export { default as CardSkeleton } from './server/CardSkeleton';
+export { default as CardListSkeleton } from './server/CardListSkeleton';
+export { default as LectureInfoSkeleton } from './server/LectureInfoSkeleton';
+export { default as ReviewCardSkeleton } from './server/ReviewCardSkeleton';
+export { default as ReviewCardListSkeleton } from './server/ReviewCardListSkeleton';
+
+// Shared Client Components
+export { default as LoginModal } from './client/LoginModal';
+export { default as Modal } from './client/Modal';
+export { default as Header } from './client/Header';
+export { default as HamburgerMenu } from './client/HamburgerMenu';
+export { default as UniversityList } from './client/UniversityList';
+export { default as DropdownUniversityList } from './client/DropdownUniversityList';
+export { default as ReactQueryProvider } from './client/ReactQueryProvider';
