+++ shared/components/server/LectureInfoSkeleton.tsx
